+++ src/aimanager/azext_aimanager/_client_factory.py
@@ -25,3 +25,7 @@ def cf_ai_manager_namespaces(cli_ctx, *_):
 
 def cf_model_deployments(cli_ctx, *_):
     return get_aimanager_client(cli_ctx).model_deployments
+
+
+def cf_ai_models(cli_ctx, *_):
+    return get_aimanager_client(cli_ctx).ai_models